CRUD users

+++ server/src/services/users.service.js
@@ -3,7 +3,7 @@ import { createUserDTO } from "../models/dtos/users.dto.js";
 import { HttpError, HTTP_STATUS } from "../utils/api.utils.js";
 import { validateUser } from "../utils/validator.js";
 
-const { usersDao, cartDao } = getDaos()
+const { usersDao, cartsDao } = getDaos()
 
 export class UsersService {
 
@@ -26,7 +26,7 @@ export class UsersService {
         return user
     }
 
-    createUser = async (payload, file) => {
+    createUser = async (payload, files) => {
 
         if (!Object.keys(payload).length) {
             throw new HttpError('Missing data for user', HTTP_STATUS.BAD_REQUEST)
@@ -34,10 +34,10 @@ export class UsersService {
 
         validateUser(payload)
 
-        const newCart = await cartDao.createCart()
+        const newCart = await cartsDao.createCart()
         payload.cart = newCart._id
 
-        const userPayloadDto = new createUserDTO(payload, file)
+        const userPayloadDto = new createUserDTO(payload, files)
 
         const newUser = await usersDao.createUser(userPayloadDto)
         return newUser
@@ -53,7 +53,7 @@ export class UsersService {
             throw new HttpError('Missing data for user', HTTP_STATUS.BAD_REQUEST)
         }
 
-        validateUser(payload)
+        // validateUser(payload)
 
         const updatedUser = await usersDao.updateUserById(id, payload)
 
@@ -64,6 +64,12 @@ export class UsersService {
 
         if (!id) {
             throw new HttpError('Missing id', HTTP_STATUS.BAD_REQUEST)
+        }
+
+        const user = await usersDao.getUserById(id)
+
+        if (!user) {
+            throw new HttpError('User not found', HTTP_STATUS.NOT_FOUND)
         }
 
         const deletedUser = await usersDao.deleteUser(id)
